+++ package.json
@@ -10,8 +10,10 @@
   },
   "dependencies": {
     "@next/font": "^13.4.7",
+    "@react-pdf/renderer": "^3.1.12",
     "next": "^13.4.9",
-    "react": "18.2.0",
-    "react-dom": "18.2.0"
+    "react": "^18.2.0",
+    "react-dom": "18.2.0",
+    "react-pdf": "^7.1.2"
   }
 }
