+++ src/app/menu/Carta/page.jsx
@@ -1,0 +1,12 @@
+import React from "react";
+import Carta from "../../components/carta";
+
+function Cartas() {
+  return (
+    <div>
+      <Carta />
+    </div>
+  );
+}
+
+export default Cartas;
